Very Simple Tagging System
right click to tag an object when it is held, this will deactivate the object and store it in a list. When we transition to night, all untagged artifacts are turned off and the tagged items are turned back on. Still need to bring them to the fire, and transition back to day